Move TLS tests to separate file
